Lazy load the video iframe
In browser that support the attribute, loading the YouTube assets is
delayed until when the iframe is about to scroll into view. This puts
the priority on loading the primary content of the page.

With the lazy loading, content from YouTube is still downloaded,
although visitors may not watch the video. A solution like Lite YouTube
Embed would only load the video when pressing play.

https://github.com/paulirish/lite-youtube-embed